feat(ui): apply a repair pairing through retoken (task 20)
Extract the persist-the-pairing-result step out of AddRemoteDialog into
apply-pairing, which now follows a repair's setToken with registry.retoken
so the live target stops sending the revoked token. The dialog drops to
290 lines and keeps only state-machine code.
